requirements.txt
streamlit>=1.32
openai>=1.0
pandas>=2.0
tabulate>=0.9
python-dotenv>=1.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
google-api-python-client>=2.0
